+++ verifier/src/main/java/com/aws/jverify/verifier/compiler/simplifications/JVerifyGhostExpressionCompiler.java
@@ -11,6 +11,7 @@ import com.sun.tools.javac.tree.TreeInfo;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.util.List;
+import java.util.Objects;
 
 public class JVerifyGhostExpressionCompiler {
     final ExpressionCompiler expressionCompiler;
@@ -116,25 +117,7 @@ public class JVerifyGhostExpressionCompiler {
                 return expressionCompiler.toExpr(args.getFirst());
             }
             case "all", "map" -> {
-                if (args.size() != 1) {
-                    throw new JavaViolationException("A %s call must have exactly one argument".formatted(methodName));
-                }
-                if (!(args.getFirst() instanceof JCTree.JCLambda lambda)) {
-                    compiler.reportError(args.getFirst(), "argumentMustBeLambda", methodName);
-                    return null;
-                }
-                var parameter = lambda.params.getFirst();
-                var paramName = parameter.getName().toString();
-                var type = compiler.translateType(parameter.type, compiler.toOrigin(parameter), null);
-                var boundVar = new BoundVar(origin, new Name(origin, paramName), type, false);
-                var body = compiler.expressionCompiler.toExpr(lambda.getBody());
-                if ("all".equals(methodName)) {
-                    return new SetComprehension(origin, List.of(boundVar), body, new IdentifierExpr(origin, paramName), null, true);
-                } else {
-                    var source = compiler.expressionCompiler.toExpr(receiver);
-                    var range = new BinaryExpr(origin, BinaryExprOpcode.In, new IdentifierExpr(origin, paramName), source);
-                    return new SetComprehension(origin, List.of(boundVar), range, body, null, true);
-                }
+                return toSetComprehension(origin, methodName, receiver, args);
             }
         }
 
@@ -147,5 +130,58 @@ public class JVerifyGhostExpressionCompiler {
         var loExpr = lo == null ? null : expressionCompiler.toExpr(lo);
         var hiExpr = hi == null ? null : expressionCompiler.toExpr(hi);
         return new SeqSelectExpr(origin, false, seqOrArrayExpr, loExpr, hiExpr, null);
+    }
+
+    private SetComprehension toSetComprehension(IOrigin origin, String methodName, JCTree.JCExpression receiver, List<JCTree.JCExpression> args) {
+        if (args.size() != 1) {
+            throw new JavaViolationException("A %s call must have exactly one argument".formatted(methodName));
+        }
+        if (!(args.getFirst() instanceof JCTree.JCLambda lambda)) {
+            compiler.reportError(args.getFirst(), "argumentMustBeLambda", methodName);
+            return null;
+        }
+        var parameter = lambda.params.getFirst();
+        var paramName = parameter.getName().toString();
+        var type = compiler.translateType(parameter.type, compiler.toOrigin(parameter), null);
+        var boundVar = new BoundVar(origin, new Name(origin, paramName), type, false);
+        var body = compiler.expressionCompiler.toExpr(lambda.getBody());
+        if ("all".equals(methodName)) {
+            return new SetComprehension(origin, List.of(boundVar), body, new IdentifierExpr(origin, paramName), null, true);
+        } else {
+            var source = compiler.expressionCompiler.toExpr(receiver);
+            var range = new BinaryExpr(origin, BinaryExprOpcode.In, new IdentifierExpr(origin, paramName), source);
+            return new SetComprehension(origin, List.of(boundVar), range, body, null, true);
+        }
+    }
+
+    /**
+     * Handles Java types from the JVerify library that are stand-ins
+     * for Dafny types defined in the JVerify prelude or built-in to Dafny.
+     */
+    public Type translateClassType(com.sun.tools.javac.code.Type.ClassType classType, IOrigin origin, boolean isNullable) {
+        var className = classType.asElement().flatName();
+        if (className.toString().equals(String.class.getName())) {
+            if (!isNullable) {
+                return new UserDefinedType(origin, new NameSegment(origin, "DString", null));
+            }
+            compiler.reportError(origin, "notSupported", "nullable String type");
+            return null;
+        }
+        if (className.toString().equals(JVerify.Sequence.class.getName())) {
+            var typeArguments = classType.getTypeArguments().stream().map(a -> compiler.translateType(a, origin)).toList();
+            if (typeArguments.stream().anyMatch(Objects::isNull)) {
+                return null;
+            }
+            return new SeqType(origin, typeArguments);
+        }
+        if (className.toString().equals(JVerify.Set.class.getName())) {
+            var arguments = classType.getTypeArguments().stream().map(a -> compiler.translateType(a, origin)).toList();
+            return new SetType(origin, arguments, true);
+        }
+        if (className.toString().equals(JVerify.Map.class.getName())) {
+            var arguments = classType.getTypeArguments().stream().map(a -> compiler.translateType(a, origin)).toList();
+            return new MapType(origin, arguments, true);
+        }
+        return null;
     }
 }
